rebase from playground

+++ srcs/render.c
@@ -23,15 +23,16 @@ static void	draw_floor(t_env *e, t_collision collision, t_coords bottom_of_wall,
     t_coords    floor_pos;
     t_coords    draw_text;
     Uint32      color_text;
+
     while (bottom_of_wall.y + 1 < WIN_H)
     {
         pixel_dist = e->op.win_h / (2.0 * bottom_of_wall.y - e->op.win_h);
-        weight_factor = (pixel_dist / collision.distance);
-        floor_pos.x = weight_factor * collision.inters.x + (1.0 - weight_factor) * player_pos.x;
+		weight_factor = (pixel_dist / collision.distance);
+		floor_pos.x = weight_factor * collision.inters.x + (1.0 - weight_factor) * player_pos.x;
         floor_pos.y = weight_factor * collision.inters.y + (1.0 - weight_factor) * player_pos.y;
         draw_text.x = (int)(floor_pos.x * PIXEL_UNIT) % derp->w;
         draw_text.y = (int)(floor_pos.y * PIXEL_UNIT) % derp->h;
-        color_text = get_pixel(derp, draw_text.x, draw_text.y, t_true);
+		color_text = get_pixel(derp, draw_text.x, draw_text.y, t_true);
         put_pixel(e->doom.surface, bottom_of_wall.x, bottom_of_wall.y, color_text);
         bottom_of_wall.y++;
     }
@@ -47,7 +48,7 @@ static void	draw_wall(t_env *e, t_collision collision, Uint32 renderer_x, double
 	y = vision_height - collision.wall_length / 2;
     y = (y < 0 ? 0 : y);
     end = vision_height + collision.wall_length / 2;
-    end = (end > e->op.win_h ? e->op.win_h : end);
+	end = (end > (int)e->op.win_h ? e->op.win_h : end);
 	draw_text.x = (int)(get_distance_between_points(collision.inters.x,
 	        collision.inters.y, collision.wall->segment.x1,
 	        collision.wall->segment.y1) * PIXEL_UNIT) % collision.wall->texture->w;
@@ -94,5 +95,5 @@ void		draw(
 	paint_floor(e, &e->op, renderer_x, e->p.vision_height);
 	skybox(e, renderer_x);
 	draw_wall(e, collision, renderer_x, e->p.vision_height);
-    draw_floor(e, collision, bottom_of_wall, e->p.pos, e->derp);
+	draw_floor(e, collision, bottom_of_wall, e->p.pos, e->derp);
 }
